+++ go.mod
@@ -3,26 +3,26 @@ module github.com/codenotary/immudb
 go 1.13
 
 require (
-	github.com/DataDog/zstd v1.4.4 // indirect
+	github.com/DataDog/zstd v1.4.5 // indirect
 	github.com/beevik/ntp v0.3.0
 	github.com/codenotary/merkletree v0.1.1
 	github.com/dgraph-io/badger/v2 v2.0.0-20200408100755-2e708d968e94
-	github.com/dgryski/go-farm v0.0.0-20191112170834-c2139c5d712b // indirect
+	github.com/dgryski/go-farm v0.0.0-20200201041132-a6ae2369ad13 // indirect
 	github.com/golang/protobuf v1.4.0
-	github.com/grpc-ecosystem/grpc-gateway v1.14.3
+	github.com/grpc-ecosystem/grpc-gateway v1.14.4
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/o1egl/paseto v1.0.0
 	github.com/pkg/errors v0.9.1 // indirect
-	github.com/prometheus/client_golang v1.2.1
+	github.com/prometheus/client_golang v1.5.1
 	github.com/rs/cors v1.7.0
-	github.com/spf13/cobra v0.0.5
-	github.com/spf13/viper v1.6.2
-	github.com/stretchr/testify v1.4.0
-	golang.org/x/crypto v0.0.0-20190308221718-c2843e01d9a2
-	golang.org/x/net v0.0.0-20200324143707-d3edc9973b7e
-	golang.org/x/sys v0.0.0-20200413165638-669c56c373c4 // indirect
+	github.com/spf13/cobra v1.0.0
+	github.com/spf13/viper v1.6.3
+	github.com/stretchr/testify v1.5.1
+	golang.org/x/crypto v0.0.0-20200423211502-4bdfaf469ed5
+	golang.org/x/net v0.0.0-20200425230154-ff2c4b7c35a0 // indirect
+	golang.org/x/sys v0.0.0-20200420163511-1957bb5e6d1f // indirect
 	golang.org/x/text v0.3.2 // indirect
-	google.golang.org/genproto v0.0.0-20200413115906-b5235f65be36
-	google.golang.org/grpc v1.28.1
+	google.golang.org/genproto v0.0.0-20200424135956-bca184e23272
+	google.golang.org/grpc v1.29.1
 	gopkg.in/yaml.v2 v2.2.8 // indirect
 )
